Server/requirements.txt
    // joblib for model dump
joblib-1.4.2
    // pandas
numpy-2.2.3
pandas-2.2.3
python-dateutil-2.9.0.post0
pytz-2025.1
six-1.17.0
tzdata-2025.1

    // pyshark
appdirs-1.4.4
lxml-5.3.1
packaging-24.2
pyshark-0.6
termcolor-2.5.0